+++ soc/arm/st_stm32/stm32h7/mpu_regions.c
@@ -5,7 +5,7 @@
  */
 
 #include <zephyr/devicetree.h>
-#include "../../common/cortex_m/arm_mpu_mem_cfg.h"
+#include <zephyr/arch/arm/aarch32/cortex_m/arm_mpu_mem_cfg.h>
 
 static const struct arm_mpu_region mpu_regions[] = {
 	MPU_REGION_ENTRY("FLASH", CONFIG_FLASH_BASE_ADDRESS,
